+++ Assets/OxGFrame/CoreFrame/Scripts/Runtime/Core/SRFrame/SRConfig/SRConfig.cs
@@ -1,11 +1,13 @@
 ﻿using MyBox;
 using System;
 using UnityEngine;
+using UnityEngine.Scripting.APIUpdating;
 
 namespace OxGFrame.CoreFrame.SRFrame
 {
+    [MovedFrom("SRSetting")]
     [Serializable]
-    public class SRSetting
+    public class SRSettings
     {
         [OverrideLabel("Exclude From Close All")]
         [Tooltip("If checked, when call CloseAll method will auto skip process")]